requirements.txt
fsspec==2024.3.1
munch==4.0.0
numpy==1.26.4
packaging==24.0
pillow==10.2.0
protobuf==5.26.1
psutil==5.9.8
PyYAML==6.0.1
scipy==1.13.0
six==1.16.0
sympy==1.12
tensorboard==2.16.2
timm==0.9.16
torch==1.13.0+cu116
torchvision==0.14.0+cu116
wheel==0.41.2
zipp==3.18.1